rag-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
python-dotenv
grpcio
grpcio-tools
